+++ Editor/TagFieldDrawer.cs
@@ -1,0 +1,52 @@
+// Copyright (c) Jerry Lee. All rights reserved. Licensed under the MIT License. See LICENSE in the
+// project root for license information.
+
+using System.Collections.Generic;
+using UniSharper;
+using UnityEditor;
+using UnityEngine;
+
+namespace UniSharperEditor
+{
+    /// <summary>
+    /// Property drawer for <see cref="TagFieldAttribute"/>.
+    /// </summary>
+    /// <seealso cref="PropertyDrawer"/>
+    [CustomPropertyDrawer(typeof(TagFieldAttribute))]
+    internal class TagFieldDrawer : PropertyDrawer
+    {
+        /// <summary>
+        /// Override this method to make your own GUI for the property.
+        /// </summary>
+        /// <param name="position">Rectangle on the screen to use for the property GUI.</param>
+        /// <param name="property">The SerializedProperty to make the custom GUI for.</param>
+        /// <param name="label">The label of this property.</param>
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            if (property.propertyType == SerializedPropertyType.String)
+            {
+                EditorGUI.BeginProperty(position, label, property);
+                
+                var tagFiledAttribute = (TagFieldAttribute)attribute;
+                if (tagFiledAttribute.UseDefaultTagFieldDrawer)
+                {
+                    property.stringValue = EditorGUI.TagField(position, label, property.stringValue);
+                }
+                else
+                {
+                    var tagList = new List<string>();
+                    tagList.AddRange(UnityEditorInternal.InternalEditorUtility.tags);
+                    var index = !string.IsNullOrEmpty(property.stringValue) ? tagList.IndexOf(property.stringValue) : -1;
+                    index = EditorGUI.Popup(position, label.text, index, tagList.ToArray());
+                    property.stringValue = index >= 0 ? tagList[index] : string.Empty;
+                }
+                
+                EditorGUI.EndProperty();
+            }
+            else
+            {
+                EditorGUI.PropertyField(position, property, label);
+            }
+        }
+    }
+}